Refine the packaging.
Previously this wasn't correctly generating things because I didn't
create a `tsconfig` which caused a `.d.ts` file not to be generated
(this is a file that editors will use to know what exists in the
package).

This commit improves that by adding a step to generate a `.d.ts` file.