Update the assertions to use `toHaveBeenCalledWith` matchers.
We would prefer to have a consistent way assert that the mocks have been
called. Because of this preference I opted to use mock object directly
and use `toEqual` in order to check the values of the calls.

This is more clear as `toHaveBeenNthCalledWith` and
`toHaveBeenCalledWith` and doesn't require type assertions. This commit
switches to using those matchers for the matchers.