docs: add some light documentation into the README
One of the things that wasn't ideal about the first version of the
`ProperLogger` for JS was that there wasn't any documentation. Users of
the package would need to either check the implementation to see how it
worked or refer to the design document.

This commit introduces some "how to use this" library documentation to
the README. I tried to separate it into "quick start" and "how does this
work".