fix: update the `ProperLogger.getlogger` to be a singleton factory
When trying to add this to the `properly-air-call-integration` (this
first application we're adding the JS `ProperLogger` to) we found that
we were tempted to pass an instance between layers. This lead me to the
conclusion that we were missing "something" because this isn't something
we need to do with the Python `ProperLogger`. The reason why we don't
need to do this with the Python `ProperLogger` is because the
`getLogger` method there keeps track of instances by name and returns
existing instance rather than creating new instances (see:
https://git.io/JUn6A)

This commit adds a similar memoized instance "thing" to the JS
`ProperLogger` by encouraging the use of the `getlogger` static method
which has been changed to return the same instance with when given the
same logger name.

This is a little different from the caching we use in the Python
`ProperLogger` because those reuse the instance of the `logger` rather
than the `ProperLogger` (the `logger` is the internal mechanism that is
wrapper by the `ProperLogger`). This means that here the "common tags"
will be shared which I feel is a desirable property.