feat: add runtime checking to `extraTags`
Ticket: https://www.pivotaltracker.com/story/show/175084603

We saw a bug a little while ago where we passed a privative value as an
extra tag. This caused logging to fail which is a problematic failure
mode.

This commit introduces a runtime check for the `extraTag` value. If a
non-object value is provided a warning will be logged that includes the
invalid tag and the original logging statement will continue without
extra tags (common tags will still be included).