feat: check that `addCommonTags` and the constructor only take valid tags
Ticket: https://www.pivotaltracker.com/story/show/175084603

We want to ensure that we only have valid tags. In the past we had an
issue with passing a non-object to one of the methods of the
`ProperLogger`, and we improved that in the previous commit by checking
the values before setting them.

This commit updates the `addCommonTags` method as well as the
constructor (which also accepts `commonTags` as it's second argument) to
have a runtime check that the pass value is a valid object.